refactor: json resume types to its specific file
