+++ src/app/resume-page/data/adapt-json-resume-project.ts
@@ -2,7 +2,7 @@ import { inject, InjectionToken } from '@angular/core'
 import { Project, Stack } from './project'
 import { DateRange } from './date-range'
 import { RELATIVIZE_PRODUCTION_URL } from '@/common/relativize-production-url'
-import { JsonResumeProject } from '../projects-section/json-resume-projects'
+import { JsonResumeProject } from '../json-resume/types'
 
 /** @visibleForTesting */
 export type AdaptJsonResumeProject = (project: JsonResumeProject) => Project
