+++ src/app/resume-page/json-resume/types.ts
@@ -5,3 +5,5 @@ export type JsonResumeWork = typeof RESUME.work
 export type JsonResumeWorkItem = JsonResumeWork[number]
 export type JsonResumeEducation = typeof RESUME.education
 export type JsonResumeEducationItem = JsonResumeEducation[number]
+export type JsonResumeProjects = typeof RESUME.projects
+export type JsonResumeProject = JsonResumeProjects[number]
